requirements.txt
argparse==1.4.0
ffmpeg-python==0.2.0
openai==0.27.7
pytube==15.0.0
